Fixed interrupts and multitasking on x64

+++ src/include/kernel/interrupts/interrupts.h
@@ -57,7 +57,7 @@ struct registers {
     uint64_t rax, rbx, rcx, rdx, rsi, rdi;
     uint64_t r8, r9, r10, r11, r12, r13, r14, r15;
     uint64_t rbp;
-    uint32_t intNum, errorCode;
+    uint64_t intNum, errorCode;
     uint64_t rip, cs, rflags, rsp, ss;
 #else
     uint32_t gs, fs, es, ds;
